typo in molecule.h, changed data type of ioff in wavefunction.h to size_t, more symmetry stuff
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@723 6a214465-e6f0-41bc-91e6-de896bb96b5e

Former-commit-id: 225cb7fee1dafd4197538f801910e537881e0471

+++ src/lib/libmints/wavefunction.cc
@@ -23,7 +23,7 @@ extern FILE *infile;
 }
 
 // Globals
-int ioff[MAX_IOFF];
+size_t ioff[MAX_IOFF];
 double df[MAX_DF];
 double bc[MAX_BC][MAX_BC];
 double fac[MAX_FAC];
@@ -92,7 +92,7 @@ void Wavefunction::initialize_singletons()
         return;
     
     ioff[0] = 0;
-    for (int i=1; i<MAX_IOFF; ++i)
+    for (size_t i=1; i<MAX_IOFF; ++i)
         ioff[i] = ioff[i-1] + i;
     
     df[0] = 1.0;
